+++ packages/tools/src/types/SegmentationStateTypes.ts
@@ -1,5 +1,9 @@
 import * as Enums from '../enums';
-import { ContourConfig, ContourSegmentationData } from './ContourTypes';
+import {
+  ContourConfig,
+  ContourRenderingConfig,
+  ContourSegmentationData,
+} from './ContourTypes';
 import type {
   LabelmapConfig,
   LabelmapRenderingConfig,
@@ -18,11 +22,8 @@ export type Color = [number, number, number, number];
 export type ColorLUT = Array<Color>;
 
 export type SegmentSpecificRepresentationConfig = {
-  [key: number]: RepresentationConfig;
+  [key: number | string]: RepresentationConfig;
 };
-/**
- * Segmentation Config
- */
 
 export type RepresentationConfig = {
   /** labelmap configuration */
@@ -126,13 +127,14 @@ export type ToolGroupSpecificLabelmapRepresentation =
 
 export type ToolGroupSpecificContourRepresentation =
   ToolGroupSpecificRepresentationState & {
-    config: LabelmapRenderingConfig;
+    config: ContourRenderingConfig;
     segmentationRepresentationSpecificConfig?: RepresentationConfig;
     segmentSpecificConfig?: SegmentSpecificRepresentationConfig;
   };
 
 export type ToolGroupSpecificRepresentation =
-  ToolGroupSpecificLabelmapRepresentation; // | other ones
+  | ToolGroupSpecificLabelmapRepresentation
+  | ToolGroupSpecificContourRepresentation;
 
 export type ToolGroupSpecificRepresentations =
   Array<ToolGroupSpecificRepresentation>;
@@ -194,10 +196,22 @@ export type ToolGroupSpecificRepresentations =
  *           colorLUTIndex: 0,
  *           visibility: true,
  *           segmentsHidden: Set(),
- *           // LabelmapRenderingConfig
+ *           // rendering config
  *           config: {
  *             "cfun",
  *             "ofun",
+ *           },
+ *           // segmentation representation specific config, has priority over the one in the outer scope
+ *           segmentationRepresentationSpecificConfig: {
+ *             LABELMAP: {
+ *               renderFill: true,
+ *             }
+ *           }
+ *           // segment specific config
+ *           segmentSpecificConfig: {
+ *             1: {
+ *              renderFill: false,
+ *              }
  *           },
  *         },
  *       ],