+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.11](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.17.10...docs@0.17.11) (2023-03-28)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+## [0.17.10](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.17.9...docs@0.17.10) (2023-03-28)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
 ## [0.17.9](https://github.com/cornerstonejs/cornerstone3D-beta/compare/docs@0.17.8...docs@0.17.9) (2023-03-28)
 
 **Note:** Version bump only for package docs